member4_model_engineer/requirements.txt
# 核心依赖
sentence-transformers>=2.2.0
# torch>=2.0.0  # 注释掉，让sentence-transformers自动安装兼容版本
transformers>=4.30.0

# 数据处理与计算
numpy>=1.24.0,<2.0  # 注意：NumPy 2.x 与 PyTorch 不兼容，必须使用 1.x
pandas>=2.0.0
scikit-learn>=1.3.0

# 配置与日志
pyyaml>=6.0
tqdm>=4.65.0

# 可选：加速检索（如需要可取消注释）
# faiss-cpu>=1.7.4

# 开发工具（可选）
# jupyter>=1.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

